Trying everything with CJS insead of ESM
