+++ src/__test__/ast.spec.ts
@@ -1,6 +1,6 @@
 import { parseAIR } from '../internal/stepper';
 
-test('== AST parsing suite', () => {
+describe('== AST parsing suite', () => {
     it('parse simple script and return ast', async function () {
         let ast = await parseAIR(`
             (call node ("service" "function") [1 2 3 arg] output)
