--- src/internal/aqua/index.d.ts
@@ -1,18 +1,0 @@
-/* tslint:disable */
-/* eslint-disable */
-/**
- * @param wasm
- * @param {string} init_user_id
- * @param {string} aqua
- * @param {string} prev_data
- * @param {string} data
- * @param {string} log_level
- * @returns {string}
- */
-export function invoke(wasm: any, init_user_id: string, aqua: string, prev_data: Uint8Array, data: Uint8Array, log_level: string): string;
-export function ast(wasm: any, script: string): string;
-export function return_current_peer_id(wasm: any, peerId: string, arg0: any): void;
-export function return_call_service_result(wasm: any, ret: string, arg0: any): void;
-export function getStringFromWasm0(wasm: any, arg1: any, arg2: any): string
-export function free(wasm: any, ptr: any, len: any): void
-
